requirements.txt
python-telegram-bot==13.12
openai==0.28.1
openai-whisper==20230918
PyPDF2==3.0.1
langchain==0.0.314
pypdf==3.16.4
chromadb==0.4.14
google-api-python-client==2.104.0
google-auth==2.23.3 
google-auth-oauthlib==1.1.0 
google-auth-httplib2==0.1.1
pytesseract==0.3.10
gtts==2.4.0
